% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_group.R
\name{arch_rolling_episodes}
\alias{arch_rolling_episodes}
\title{@rdname  episode_group}
\usage{
arch_rolling_episodes(x, strata = NULL, case_length,
  recurrence_length = NULL, from_last = FALSE,
  overlap_method = c("across", "inbetween", "aligns_start", "aligns_end",
  "chain"), deduplicate = FALSE, display = TRUE)
}
\description{
@rdname  episode_group
}
