% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bys_funcs_legacy.R
\name{bys_funcs_legacy}
\alias{bys_funcs_legacy}
\alias{bys_count_legacy}
\alias{bys_rank_legacy}
\alias{bys_position_legacy}
\alias{bys_val_legacy}
\alias{bys_nval_legacy}
\alias{bys_min_legacy}
\alias{bys_max_legacy}
\alias{bys_sum_legacy}
\alias{bys_prod_legacy}
\alias{bys_cummin_legacy}
\alias{bys_cummax_legacy}
\alias{bys_cumsum_legacy}
\alias{bys_cumprod_legacy}
\alias{bys_lag_legacy}
\alias{bys_lead_legacy}
\title{Vectorised approach to group operations.}
\usage{
bys_count_legacy(by, unique.var = NULL)

bys_rank_legacy(..., by = NULL, from_last = FALSE)

bys_position_legacy(val, by = NULL, from_last = FALSE, ordered = TRUE)

bys_val_legacy(..., val, by = NULL, from_last = FALSE)

bys_nval_legacy(..., val, by = NULL, from_last = FALSE, n = 1, nmax = FALSE)

bys_min_legacy(val, by = NULL, na.rm = TRUE)

bys_max_legacy(val, by = NULL, na.rm = TRUE)

bys_sum_legacy(val, by = NULL, na.rm = TRUE, cumulative = FALSE)

bys_prod_legacy(val, by = NULL, na.rm = TRUE, cumulative = FALSE)

bys_cummin_legacy(val, by = NULL, na.rm = TRUE)

bys_cummax_legacy(val, by = NULL, na.rm = FALSE)

bys_cumsum_legacy(val, by = NULL, na.rm = TRUE)

bys_cumprod_legacy(val, by = NULL, na.rm = TRUE)

bys_lag_legacy(val, by = NULL, n = 1)

bys_lead_legacy(val, by = NULL, n = 1)
}
\arguments{
\item{by}{\code{[atomic]}. Groups.}

\item{...}{\code{[atomic]}. Sort levels.}

\item{from_last}{\code{[logical]} Sort order - \code{TRUE} (descending) or \code{FALSE} (ascending).}

\item{val}{\code{[atomic]}. Value.}

\item{ordered}{If \code{TRUE}, the output is sequential.}

\item{n}{\code{[integer]} Position.}

\item{nmax}{\code{[logical]} If \code{TRUE}, use \code{length([by])} when \code{n} is greater than the number of records in a group.}

\item{na.rm}{If \code{TRUE}, remove \code{NA} values}
}
\value{
\code{[atomic]}
}
\description{
Vectorised approach to group operations.
}
\examples{
x <- data.frame(
  group = c(2, 2, 1, 2, 1, 1, 1, 2, 1, 1),
  value = c(13, 14, 20, 9, 2, 1, 8, 18, 3, 17))

bys_count_legacy(x$group)
bys_position_legacy(x$value, by = x$group, from_last = TRUE)
bys_rank_legacy(by = x$group, val = x$value, from_last = TRUE)
bys_val_legacy(x$value, by = x$group, val = x$value, from_last = TRUE)
bys_nval_legacy(x$value, by = x$group, val = x$value, from_last = TRUE, n = 2)
bys_min_legacy(by = x$group, val = x$value)
bys_max_legacy(by = x$group, val = x$value)
bys_sum_legacy(by = x$group, val = x$value)
bys_prod_legacy(by = x$group, val = x$value)
bys_cummin_legacy(by = x$group, val = x$value)
bys_cummax_legacy(by = x$group, val = x$value)
bys_cumsum_legacy(by = x$group, val = x$value)
bys_cumprod_legacy(by = x$group, val = x$value)
bys_lag_legacy(by = x$group, val = x$value)
bys_lead_legacy(by = x$group, val = x$value)
}
