% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi.R
\name{combi}
\alias{combi}
\title{Vector combinations}
\usage{
combi(..., ordered = FALSE)
}
\arguments{
\item{...}{\code{[atomic]}.}

\item{ordered}{\code{[logical]}. If \code{TRUE}, the output is sequential.}
}
\value{
\code{numeric}
}
\description{
Numeric codes for unique combination of vectors.
}
\examples{
x <- c("A", "B", "A", "C", "B", "B")
y <- c("X", "X", "Z", "Z", "X", "Z")
combi(x, y)
}
