% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_funcs.R
\name{concat_code}
\alias{concat_code}
\title{Concatenate strings}
\usage{
concat_code(...)
}
\arguments{
\item{...}{\code{[atomic]}}
}
\value{
\code{numeric}
}
\description{
Numeric codes for unique concatenation of strings.
}
\details{
For large datasets, \code{concat_code} provides a faster solution to concatenating variables compared to \code{paste0} and \code{paste}.
}
\examples{
x <- c("A", "B", "A", "C", "B", "B")
y <- c("X", "X", "Z", "Z", "X", "Z")
concat_code(x, y)

# The above is equivalent but quicker than efficiently than this.
z <- paste0(y, "-", x)
z <- match(z, z)
z
}
