% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epid.R
\docType{class}
\name{epid-class}
\alias{epid-class}
\alias{show,epid-method}
\alias{rep,epid-method}
\alias{[,epid-method}
\alias{[[,epid-method}
\alias{$,epid-method}
\title{S4 objects containing the results of \code{episode_grouping()}, \code{fixed_epiosdes()} and \code{rolling_episode()}}
\usage{
\S4method{show}{epid}(object)

\S4method{rep}{epid}(x, ...)

\S4method{[}{epid}(x, i, j, ..., drop = TRUE)

\S4method{[[}{epid}(x, i, j, ..., exact = TRUE)

\S4method{$}{epid}(x, name)
}
\arguments{
\item{object}{object}

\item{x}{x}

\item{...}{...}

\item{i}{i}

\item{j}{j}

\item{drop}{drop}

\item{exact}{exact}

\item{name}{slot name}
}
\description{
S4 objects containing the results of \code{episode_grouping()}, \code{fixed_epiosdes()} and \code{rolling_episode()}
}
\section{Slots}{

\describe{
\item{\code{start}}{Start of the number line}

\item{\code{id}}{Unique \code{numeric} ID. Providing this is optional.}

\item{\code{gid}}{Unique \code{numeric} Group ID. Providing this is optional.}

\item{\code{.Data}}{Length/with and direction of the \code{number_line} object.}
}}

