% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epid.R
\name{epid}
\alias{epid}
\alias{unique.epid}
\alias{format.epid}
\title{Number line objects}
\usage{
epid(x)

\method{unique}{epid}(x, ...)

\method{format}{epid}(x, ...)
}
\arguments{
\item{...}{arguments for particular methods}
}
\description{
Number line objects
}
