% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_funcs.R
\name{finite_check}
\alias{finite_check}
\alias{conv_funcs}
\title{Check for numeric values that are not finite}
\usage{
finite_check(x)
}
\arguments{
\item{x}{\code{numeric} object or can be coerced to \code{numeric} object}
}
\value{
\code{character} object
}
\description{
Check for numeric values that are not finite
}
