% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_datasets.R
\docType{data}
\name{hospital_admissions}
\alias{hospital_admissions}
\title{Dummy data for \code{diyar}}
\format{tibble}
\usage{
data(hospital_admissions)
}
\description{
Dated hospital admissions and discharges
}
\examples{
data(hospital_admissions)
}
\keyword{datasets}
