% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_datasets.R
\docType{data}
\name{infections}
\alias{infections}
\title{Dummy data for \code{diyar}}
\format{tibble}
\usage{
data(infections)
}
\description{
Dated reports of bacterial infections
}
\examples{
data(infections)
}
\keyword{datasets}
