% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_group.R
\name{links_probabilistic}
\alias{links_probabilistic}
\alias{fetch_thresholds}
\title{Probabilistic record linkage}
\usage{
links_probabilistic(
  attribute,
  blocking_attribute = NULL,
  cmp_func = diyar::exact_match,
  cmp_threshold = 0.95,
  probabilistic = TRUE,
  m_prob = 0.95,
  weight_threshold = 1,
  ...
)

fetch_thresholds(attribute, m_prob = 0.99)
}
\arguments{
\item{attribute}{\code{list} of attributes to compare.}

\item{blocking_attribute}{Subsets of the dataset.}

\item{cmp_func}{\code{list} of functions to used as string comparators for each \code{attribute}. See \code{Details}.}

\item{cmp_threshold}{\code{list} of real numbers. These are thresholds for what's considered a match. See \code{Details}.}

\item{probabilistic}{if \code{TRUE}, matching weights are calculated using XXX and XXX method for XXX.}

\item{m_prob}{m-probability for each attribute.}

\item{weight_threshold}{Minimum threshold for linked records. See \code{Details}.}

\item{...}{Arguments passed to \bold{\code{links}}}
}
\value{
\code{\link[=pid-class]{pid}} or \code{list} (\code{\link[=pid-class]{pid}} and \code{ggplot}) object
}
\description{
A specific use case of \code{links} to achieve probabilistic record linkage.
}
\details{
XXXXXX
}
\seealso{
\code{\link{links}},  \code{\link{episodes}}, \code{\link{partitions}}, \code{\link{predefined_tests}} and \code{\link{sub_criteria}}
}
