% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_funcs.R
\name{listr}
\alias{listr}
\title{Grammatical lists.}
\usage{
listr(x, sep = ", ", conj = " and", lim = Inf)
}
\arguments{
\item{x}{\code{atomic} vector.}

\item{sep}{Separator.}

\item{conj}{Final separator.}

\item{lim}{Elements to include in the list. Other elements are abbreviated to \code{" ..."}.}
}
\value{
\code{character}.
}
\description{
A convenience function to format \code{atomic} vectors as a written list.
}
\examples{
listr(1:5)
listr(1:5, sep = "; ")
listr(1:5, sep = "; ", conj = " and")
listr(1:5, sep = "; ", conj = " and", lim = 2)

}
