% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_datasets.R
\docType{data}
\name{missing_staff_id}
\alias{missing_staff_id}
\title{Dummy data for \code{diyar}}
\format{tibble}
\usage{
data(missing_staff_id)
}
\description{
Dummy staff records with missing staff identifiers
}
\examples{
data(missing_staff_id)
}
\keyword{datasets}
