% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{pane-class}
\alias{pane-class}
\alias{as.pane}
\alias{format.pane}
\alias{unique.pane}
\alias{summary.pane}
\alias{show,pane-method}
\alias{rep,pane-method}
\alias{[,pane-method}
\alias{[[,pane-method}
\alias{c,pane-method}
\title{\code{pane} object}
\usage{
as.pane(x)

\method{format}{pane}(x, ...)

\method{unique}{pane}(x, ...)

\method{summary}{pane}(x, ...)

\S4method{show}{pane}(object)

\S4method{rep}{pane}(x, ...)

\S4method{[}{pane}(x, i, j, ..., drop = TRUE)

\S4method{[[}{pane}(x, i, j, ..., exact = TRUE)

\S4method{c}{pane}(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{...}

\item{object}{object}

\item{i}{i}

\item{j}{j}

\item{drop}{drop}

\item{exact}{exact}
}
\description{
S4 objects to store the results of \code{\link{episodes}}
}
