% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panes.R
\name{panes}
\alias{panes}
\title{Track episodes for case definitions and record deduplication - Panes.}
\usage{
panes(
  date,
  window = Inf,
  windows_total = 1,
  separate = FALSE,
  sn = NULL,
  strata = NULL,
  data_links = "ANY",
  custom_sort = NULL,
  group_stats = FALSE,
  display = "none",
  data_source = NULL,
  by = NULL,
  length.out = NULL,
  fill = TRUE,
  schema = "none"
)
}
\arguments{
\item{date}{Event date (\code{date}, \code{datetime} or \code{numeric}) or period (\code{\link{number_line}}).}

\item{window}{Numeric or time interval supplied as \code{number_line} objects.}

\item{windows_total}{Minimum number of overlapping windows in an in an episode. See \code{details}}

\item{separate}{If \code{TRUE}, events under one window are considered separate episodes from those in another. The default is \code{FALSE}.}

\item{sn}{Unique numerical record identifier. Useful for creating familiar episode identifiers.}

\item{strata}{Subsets. Episodes are tracked separately within each subset. \code{\link{links}} is useful for creating these.}

\item{data_links}{A set of \code{data_sources} required in each episode. A \code{strata} without records from these data sources will be skipped, and episodes without these will be unlinked. See \code{Details}.}

\item{custom_sort}{Preferential order for selecting index (\code{"index"}) events.}

\item{group_stats}{If \code{TRUE} (default), episode-specific information like episode start and endpoints are returned. See \code{Value}.}

\item{display}{The messages printed on screen. Options are; \code{"none"} (default) or and \code{"stats"} for a summary of thes process.}

\item{data_source}{Unique data source identifier. Useful when the dataset has data from multiple sources.}
}
\value{


\code{\link[=epid-class]{epid}} objects or \code{data.frame} if \code{to_s4} is \code{FALSE}

\itemize{
\item \code{sn} - unique record identifier as provided (or generated)
\item \code{epid | .Data} - unique episode identifier
\item \code{wind_id} - unique window identifier
\item \code{wind_nm} - type of window i.e. "Index" or "Skipped"
\item \code{case_nm} - record type in regards to index assignment
\item \code{dist_from_wind} - duration of each event from its window's reference event
\item \code{dist_from_epid} - duration of each event from its episode's reference event
\item \code{epid_dataset} - data sources in each episode
\item \code{epid_interval} - episode start and end dates. A \code{\link{number_line}} object.
\item \code{epid_length} - the difference between episode start and end dates (\code{difftime}).
\item \code{epid_total} - number of records in each episode
\item \code{iteration} - iteration of the process when each event was tracked to its episode.
}
}
\description{
Distribute events into panes defined by time or numerical intervals.
}
\details{
Panes are events that overlaps with particular time or numerical intervals (\code{window}).
Events that overlap with one window are considered part of the same pane and different from events that overlap with another window.

\code{Overlapping windows are merged}
\code{events across two windows are assigned to the last one listed}

Unlike \code{\link{episodes}}, panes are not tracked in a sequential order.
However, by default, the earliest event or smallest record is selected as the \code{"Index"} event.
\code{custom_sort} provies some flexibility with regards to which record is taken as the index assignment.

\bold{\code{panes()}} will categorise records into 3 types of events;

\itemize{
\item \code{"Index"} - Index event/record of the pane.
\item \code{"Duplicate_I"} - Duplicate of the index case.
\item \code{"Skipped"} - Those skipped from the pane distribution process.
}

\code{data_source} - including this populates the \code{epid_dataset} slot. See \code{Value}.

\code{data_links} should be a \code{list} of \code{atomic} vectors with every element named \code{"l"} (links) or \code{"g"} (groups).
\itemize{
\item \code{"l"} - Episodes with records from every listed data source will be retained.
\item \code{"g"} - Episodes with records from any listed data source will be retained.
}
\code{data_links} is useful for skipping events or records that are not required to minimise processing time.

See \code{vignette("panes")} for more information.
}
\examples{
library(diyar)

events <- c(30, 2, 11, 10, 100)
windows <- number_line(c(1, 9, 25), c(3, 12, 35))

events
panes(date = events, window = windows, separate = T)
panes(date = events, window = windows, separate = F)

panes(date = events, window = windows, separate = T, windows_total = 3)
panes(date = events, window = windows, separate = F, windows_total = 4)

}
\seealso{
\code{\link{episodes}}, \code{\link{links}}, \code{\link{overlaps}} and \code{\link{number_line}}
}
