% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panes.R
\name{partitions}
\alias{partitions}
\title{Separate events into time intervals.}
\usage{
partitions(
  date,
  window = number_line(0, Inf),
  windows_total = 1,
  separate = FALSE,
  sn = NULL,
  strata = NULL,
  data_links = "ANY",
  custom_sort = NULL,
  group_stats = FALSE,
  data_source = NULL,
  by = NULL,
  length.out = NULL,
  fill = TRUE,
  schema = "none",
  ...
)
}
\arguments{
\item{date}{Event date (\code{date}, \code{datetime} or \code{numeric}) or period (\code{\link{number_line}}).}

\item{window}{Numeric or time intervals. Supplied as \code{number_line} objects.}

\item{windows_total}{Minimum number of matched \code{windows} required for pane. See \code{details}}

\item{separate}{If \code{TRUE}, events matched to different \code{windows} are not linked.}

\item{sn}{Unique numerical record identifier. Useful for creating familiar pane identifiers.}

\item{strata}{Subsets of the dataset. Panes are created separately for each \code{strata}. See \code{\link{links}}.}

\item{data_links}{A set of \code{data_sources} required for pane. A \code{strata} without records from these \code{data_sources} are skipped. See \code{Details}.}

\item{custom_sort}{Preferred order for selecting \code{"index"} events.}

\item{group_stats}{If \code{TRUE} (default), the returned \code{pane} object will include pane-specific information like panes' start and endpoints.}

\item{data_source}{Unique data source identifier. Useful when the dataset has data from multiple sources.}
}
\value{


\code{\link[=pane-class]{pane}} objects or \code{data.frame} if \code{to_s4} is \code{FALSE}

\itemize{
\item \code{sn} - unique record identifier as provided (or generated)
\item \code{pane | .Data} - unique episode identifier
\item \code{case_nm} - record type in regards to index assignment
\item \code{window_list} - record type in regards to index assignment
\item \code{dist_wind_index} - duration of each event from its window's reference event
\item \code{dist_pane_index} - duration of each event from its episode's reference event
\item \code{pane_dataset} - data sources in each episode
\item \code{pane_interval} - episode start and end dates. A \code{\link{number_line}} object.
\item \code{pane_length} - the difference between episode start and end dates (\code{difftime}).
\item \code{pane_total} - number of records in each episode
\item \code{iteration} - iteration of the process when each event was tracked to its episode.
}
}
\description{
Distribute events into partitions defined by time or numerical intervals.
Records in each partition are assigned a unique identifier with relevant group data - \code{pane} objects.
}
\details{
A pane consists of events that match to a specific time or numerical intervals (\code{window}).

Each \code{Windows} must cover a separate interval. Overlapping \code{Windows} are merged before events are partitioned.
Events that occur during two \code{windows} are assigned to the last one listed.

Alternatively, you can create \code{windows} by splitting the period between the earliest and most recent time points into equal parts (\code{length.out}), or a sequence of intervals with fixed widths (\code{by}).

By default, the earliest event is taken as the \code{"Index"} event.
An alternative can be chosen with \code{custom_sort}.

\bold{\code{partitions()}} will categorise records into 3 types;

\itemize{
\item \code{"Index"} - Index event/record of the pane.
\item \code{"Duplicate_I"} - Duplicate of the \code{"Index"} case.
\item \code{"Skipped"} - Records that are not assigned to a pane.
}

If \code{data_source} is provided, \code{pane_dataset} is added to the \code{pane} object. See \code{Value}.

\code{data_links} must be a \code{list} of \code{atomic} vectors, with every element named \code{"l"} (links) or \code{"g"} (groups).

\itemize{
\item \code{"l"} - only panes with records from every listed \code{data_source} will be retained.
\item \code{"g"} - only panes with records from any listed data source will be retained.
}

Assigning \code{NA} to \code{data_links} will exclude an event from being partitioned.

See \code{vignette("partitions")} for more information.
}
\examples{
events <- c(30, 2, 11, 10, 100)
windows <- number_line(c(1, 9, 25), c(3, 12, 35))

events
partitions(date = events, window = windows, separate = TRUE)
partitions(date = events, window = windows, separate = FALSE)

partitions(date = events, window = windows, separate = TRUE, windows_total = 3)
partitions(date = events, window = windows, separate = FALSE, windows_total = 4)

}
\seealso{
\code{\link[=pane-class]{pane}}, \code{\link{number_line_sequence}}, \code{\link{episodes}}, \code{\link{links}}, \code{\link{overlaps}} and \code{\link{number_line}}
}
