% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_datasets.R
\docType{data}
\name{patient_list}
\alias{patient_list}
\title{Dummy data for \code{diyar}}
\format{data.frame}
\usage{
data(patient_list)
}
\description{
Patient list with some missing data
}
\examples{
data(patient_list)
}
\keyword{datasets}
