% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{pid-class}
\alias{pid-class}
\alias{as.pid}
\alias{format.pid}
\alias{unique.pid}
\alias{summary.pid}
\alias{show,pid-method}
\alias{rep,pid-method}
\alias{[,pid-method}
\alias{[[,pid-method}
\alias{c,pid-method}
\title{\code{pid} objects}
\usage{
as.pid(x, ...)

\method{format}{pid}(x, ...)

\method{unique}{pid}(x, ...)

\method{summary}{pid}(x, ...)

\S4method{show}{pid}(object)

\S4method{rep}{pid}(x, ...)

\S4method{[}{pid}(x, i, j, ..., drop = TRUE)

\S4method{[[}{pid}(x, i, j, ..., exact = TRUE)

\S4method{c}{pid}(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{...}

\item{object}{object}

\item{i}{i}

\item{j}{j}

\item{drop}{drop}

\item{exact}{exact}
}
\description{
S4 objects to store the results of \code{\link{record_group}}
}
