% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plots}
\alias{plots}
\alias{plot_pids}
\title{Plots}
\usage{
plot_pids(pids, title = NULL, plot_labels = FALSE)
}
\arguments{
\item{pids}{\code{pid} object}

\item{title}{Title of \code{ggplot} object}
}
\value{
\code{ggplot}
}
\description{
Plots
}
\details{

}
