% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_datasets.R
\docType{data}
\name{staff_records}
\alias{staff_records}
\title{Dummy data for \code{diyar}}
\format{tibble}
\usage{
data(staff_records)
}
\description{
Dummy staff record with some missing data
}
\examples{
data(staff_records)
}
\keyword{datasets}
