% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_criteria.R
\name{sub_criteria}
\alias{sub_criteria}
\alias{attrs}
\alias{eval_sub_criteria}
\alias{print.sub_criteria}
\alias{eval_sub_criteria.sub_criteria}
\title{Nested linkage criteria}
\usage{
sub_criteria(
  ...,
  match_funcs = diyar::exact_match,
  equal_funcs = diyar::exact_match,
  operator = "or"
)

attrs(..., .obj = NULL)

eval_sub_criteria(x, ...)

\method{print}{sub_criteria}(object, ...)

\method{eval_sub_criteria}{sub_criteria}(
  x,
  x_pos = seq_len(max(attr_eval(x))),
  y_pos = rep(1L, length(x_pos)),
  check_duplicates = TRUE,
  ...
)
}
\arguments{
\item{...}{\code{[atomic]} Attributes passed to or \code{eval_sub_criteria()} or \code{eval_sub_criteria()}

Arguments passed to methods for \code{eval_sub_criteria()}}

\item{match_funcs}{\code{[function]}. User defined logical test for matches.}

\item{equal_funcs}{\code{[function]}. User defined logical test for identical record sets (all attributes of the same record).}

\item{operator}{\code{[character]}. Options are \code{"and"} or \code{"or"}.}

\item{.obj}{\code{[data.frame|list]}. Attributes.}

\item{x_pos}{\code{[integer]}. Index of one half of a record pair.}

\item{y_pos}{\code{[integer]}. Index of one half of a record pair.}

\item{check_duplicates}{\code{[logical]}. If \code{FALSE}, does not check duplicate values. The result of the initial check will be recycled.}
}
\value{
\code{\link{sub_criteria}}
}
\description{
Nested matching criteria for each iteration of \bold{\code{\link{links}}} and \bold{\code{\link{episodes}}}.
}
\details{
\bold{\code{sub_criteria()}} - Creates a \code{sub_criteria} object - a nested criteria evaluated at each iteration of \bold{\code{links}} or \bold{\code{episodes}}.
A \code{sub_criteria} object contains attributes to be compared,
logical tests for the comparisons (see \bold{\code{\link{predefined_tests}}} for examples) and
another set of logical tests to determine identical records.

\bold{\code{attrs()}} - Create as \code{d_attribute} object - a collection of objects that can be passed as a single attribute to \bold{\code{sub_criteria()}}.

\bold{\code{eval_sub_criteria()}} - Evaluates a \code{sub_criteria} object.

At each iteration of \code{\link{links}} or \code{\link{episodes}}, record-pairs are created from each attribute of the \code{sub_criteria} object.
\code{eval_sub_criteria()} evaluates each record-pair using the \code{match_funcs} and \code{equal_funcs} functions of the \code{sub_criteria} object.
See \bold{\code{\link{predefined_tests}}} for examples of \code{match_funcs} and \code{equal_funcs}.
\code{attrs()} is useful when the criteria requires an interaction between the multiple attributes. For example, attribute 1 + attribute 2 > attribute 3.

Every attribute, including those in \code{attrs()}, must have the same length or a length of 1.
}
\examples{
# Attributes
attr_1 <- c(30, 28, 40, 25, 25, 29, 27)
attr_2 <- c("M", "F", "U", "M", "F", "U", "M")

# Creating a nested criteria
## Example 1 - A maximum difference of 10 in attribute 1
s_cri1 <- sub_criteria(attr_1, match_funcs = range_match)
s_cri1

# Evaluating the nested criteria
## .. at the 1st iteration of linkage
eval_sub_criteria(s_cri1)
## .. at the 2nd iteration of linkage
x_pos_val <- seq_len(max(attr_eval(s_cri1)))
eval_sub_criteria(s_cri1,
                  x_pos = x_pos_val,
                  y_pos = rep(2, length(x_pos_val)))

## Example 2 - `s_cri1` AND an exact match on attribue 2
s_cri2 <- sub_criteria(
  s_cri1,
  sub_criteria(attr_2, match_funcs = exact_match),
  operator = "and")
s_cri2

## Example 3 - `s_cri1` OR an exact match on attribue 2
s_cri3 <- sub_criteria(
  s_cri1,
  sub_criteria(attr_2, match_funcs = exact_match),
  operator = "or")
s_cri3

# Evaluating the nested criteria
eval_sub_criteria(s_cri2)
eval_sub_criteria(s_cri3)

# Alternatively, using `attr()`
AND_func <- function(x, y) range_match(x$a1, y$a1) & x$a2 == y$a2
OR_func <- function(x, y) range_match(x$a1, y$a1) | x$a2 == y$a2

## Creating the nested criteria
s_cri2b <- sub_criteria(attrs(.obj = list(a1 = attr_1, a2 = attr_2)),
                        match_funcs = AND_func)
s_cri3b <- sub_criteria(attrs(.obj = list(a1 = attr_1, a2 = attr_2)),
                        match_funcs = OR_func)

# Evaluate the nested criteria
eval_sub_criteria(s_cri2b)
eval_sub_criteria(s_cri3b)

}
\seealso{
\code{\link{predefined_tests}}; \code{\link{links}}; \code{\link{episodes}}; \code{\link{eval_sub_criteria}}
}
