% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_df.R
\name{transfer}
\alias{transfer}
\alias{to_epid}
\alias{to_df}
\title{Converting}
\usage{
to_epid(df)

to_df(s4)
}
\arguments{
\item{df}{\code{df}}

\item{s4}{s4}
}
\value{
\code{epid} object class

\code{data.frame} object class
}
\description{
Convert
}
