% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_group.R
\name{user_defined_tests}
\alias{user_defined_tests}
\alias{range_matching}
\title{User defined tests}
\usage{
range_matching(x, y, range = 10)
}
\description{
Collection of user defined tests for \code{\bold{\link{sub_criteria}}}
}
\examples{
range_matching(10, 22, range = 6)
range_matching(10, 16, range = 6)
range_matching(16, 10, range = 6)
}
