% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_funcs.R
\name{combi}
\alias{combi}
\title{Vector combinations}
\usage{
combi(...)
}
\arguments{
\item{...}{\code{[atomic]}}
}
\value{
\code{numeric}
}
\description{
Numeric codes for unique combination of vectors.
}
\examples{
x <- c("A", "B", "A", "C", "B", "B")
y <- c("X", "X", "Z", "Z", "X", "Z")
combi(x, y)

# The code above is equivalent to but quicker than the one below.
z <- paste0(y, "-", x)
z <- match(z, z)
z
}
