% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_group.R
\name{custom_sort}
\alias{custom_sort}
\title{Nested sorting}
\usage{
custom_sort(..., decreasing = FALSE)
}
\arguments{
\item{...}{Sequence of \code{atomic} vectors. Passed to \bold{link\code{order}}.}

\item{decreasing}{Sort order. Passed to \bold{link\code{order}}.}
}
\value{
\code{numeric} sort order.
}
\description{
Returns a sort order after sorting by a vector within another vector.
}
\examples{

a <- c(1, 1, 1, 2, 2)
b <- c(2, 3, 2, 1, 1)

custom_sort(a, b)
custom_sort(b, a)

}
