% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{d_report}
\alias{d_report}
\alias{plot.d_report}
\alias{as.list.d_report}
\alias{as.data.frame.d_report}
\title{d_report}
\usage{
\method{plot}{d_report}(x, ...)

\method{as.list}{d_report}(x, ...)

\method{as.data.frame}{d_report}(x, ...)
}
\arguments{
\item{x}{\code{[d_report]}.}

\item{...}{Arguments passed to other methods}
}
\description{
d_report
}
