% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{epid-class}
\alias{epid-class}
\alias{format.epid}
\alias{show,epid-method}
\alias{rep,epid-method}
\alias{[,epid-method}
\alias{[[,epid-method}
\title{S4 classes in \code{diyar}}
\usage{
\method{format}{epid}(x, ...)

\S4method{show}{epid}(object)

\S4method{rep}{epid}(x, ...)

\S4method{[}{epid}(x, i, j, ..., drop = TRUE)

\S4method{[[}{epid}(x, i, j, ..., exact = TRUE)
}
\arguments{
\item{x}{x}

\item{...}{...}

\item{object}{object}

\item{i}{i}

\item{j}{j}

\item{drop}{drop}

\item{exact}{exact}
}
\description{
S4 classes in \code{diyar}
}
