% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{epid-class}
\alias{epid-class}
\alias{as.epid}
\alias{format.epid}
\alias{unique.epid}
\alias{summary.epid}
\alias{show,epid-method}
\alias{rep,epid-method}
\alias{[,epid-method}
\alias{[[,epid-method}
\alias{c,epid-method}
\title{\code{epid} object}
\usage{
as.epid(x)

\method{format}{epid}(x, ...)

\method{unique}{epid}(x, ...)

\method{summary}{epid}(x, ...)

\S4method{show}{epid}(object)

\S4method{rep}{epid}(x, ...)

\S4method{[}{epid}(x, i, j, ..., drop = TRUE)

\S4method{[[}{epid}(x, i, j, ..., exact = TRUE)

\S4method{c}{epid}(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{...}

\item{object}{object}

\item{i}{i}

\item{j}{j}

\item{drop}{drop}

\item{exact}{exact}
}
\description{
S4 objects to store the results of \code{\link{episodes}}
}
