% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\name{episodes_wf_splits}
\alias{episodes_wf_splits}
\title{XXXXXXXXXXXXXXXXXXXXXX.}
\usage{
episodes_wf_splits(..., duplicates_recovered = "ALL", reframe = FALSE)
}
\arguments{
\item{...}{Arguments passed to \code{\link{episodes}}.}

\item{duplicates_recovered}{XXXXXXXXXXXXXXXXXXXXXXXXXX}

\item{reframe}{XXXXXXXXXXXXXXXXXXXXXXXXXX}
}
\value{
\code{\link[=epid-class]{epid}}; \code{list}
}
\description{
XXXXXXXXXXXXXXXXXXXXXXXXXXXXX.
}
\details{
\bold{\code{episodes_wf_splits()}} is a wrapper function of \bold{\code{episodes()}} which strips or re-frames the dataset to
the minimum number of records required to implement a case definition.
This leads to the same outcome as would \code{\link{episodes}} but with benefit of shorter processing times.
It is recommended for larger datasets with duplicate events.

If a \code{\link{sub_criteria}} is passed to \code{case_sub_criteria} or \code{recurrence_sub_criteria},
duplicate records in each attribute can either be stripped or reframed using the \code{reframe} argument. See \code{Examples}.

See \code{vignette("episodes")} for further details.
}
\examples{
10,000 duplicate records of 20 events
dates <- seq(from = as.Date("2019-04-01"), to = as.Date("2019-04-20"), by = 1)
dates <- rep(dates, 10000)

# `episodes_wf_splits` will generally take less more time
system.time(
  ep1 <- episodes(dates, 1)
)

system.time(
  ep2 <- episodes_wf_splits(dates, 1)
)

# Results will be identical
all(ep1 == ep2)

# Split vs Reframed `sub_criteria`
 # Duplicate values for an attribute - color
 attr <- colours()[1:5]
 attr <- rep(attr, 3)
 # A running number as `sub_criteria`
 sub_cri <- sub_criteria(1:length(attr))
 # Having `reframe` as `TRUE` or `FALSE` will either;
   # split the `sub_criteria` OR
   split_scri <- diyar:::sp_scri(sub_cri, !duplicated(attr))
   # reframed the `sub_criteria`
   reframed_scri <- diyar:::rf_scri(sub_cri, attr)

 # The new attributes
 attr_eval(split_scri, identity, simplify = FALSE)
 attr_eval(reframed_scri, identity, simplify = FALSE)
}
\seealso{
\code{\link{episodes}}; \code{\link{sub_criteria}}
}
