% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_sub_criteria.R
\name{eval_sub_criteria}
\alias{eval_sub_criteria}
\alias{eval_sub_criteria.sub_criteria}
\title{Evaluate a \code{\link{sub_criteria}}.}
\usage{
eval_sub_criteria(x, ...)

\method{eval_sub_criteria}{sub_criteria}(
  x,
  strata = seq_len(max(attr_eval(x))),
  index_record = c(TRUE, rep(FALSE, length(strata) - 1)),
  sn = seq_len(length(strata)),
  check_duplicates = TRUE,
  ...
)
}
\arguments{
\item{x}{\code{[\link{sub_criteria}]}.}

\item{...}{Arguments passed to methods.}

\item{strata}{\code{[integer]}. Subsets of the dataset}

\item{index_record}{\code{[logical]}. Index record (\code{y}) being compared against other records (\code{x}).
See (\bold{\code{\link{predefined_tests}}}).}

\item{sn}{\code{[integer]} Unique index for each record.}

\item{check_duplicates}{\code{[logical]}. If \code{FALSE}, does not check duplicate values. The result of the initial check will be recycled.}
}
\value{
\code{logical}; \code{list}
}
\description{
Evaluate a \code{\link{sub_criteria}}.
}
\examples{
# Consider three records
strata <- rep(1, 3)
index_record <- c(TRUE, FALSE, FALSE)
sn <- 1:3

# Test for a match in either attribute
sub_cri_1 <- sub_criteria(c(1, 1, 0), c(2, 1, 2))
eval_sub_criteria(sub_cri_1, strata, index_record, sn)

# Test for a match in both attributes
sub_cri_2 <- sub_criteria(c(1, 1, 0), c(2, 1, 2), operator = "and")
eval_sub_criteria(sub_cri_2, strata, index_record, sn)
}
\seealso{
\code{\link{sub_criteria}}; \code{\link{reframe}}
}
