% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_funcs.R
\name{finite_check}
\alias{finite_check}
\alias{enq_vr}
\alias{fmt}
\alias{listr}
\alias{duplicates_check}
\alias{logicals_check}
\title{Convenience functions}
\usage{
finite_check(x)

enq_vr(x)

fmt(x)

listr(x)

duplicates_check(x)

logicals_check(x)
}
\arguments{
\item{x}{x}
}
\description{
Convenience functions
}
