% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_group.R
\name{links}
\alias{links}
\alias{record_group}
\title{Multistage deterministic record linkage}
\usage{
links(
  criteria,
  sub_criteria = NULL,
  sn = NULL,
  strata = NULL,
  data_source = NULL,
  data_links = "ANY",
  display = "none",
  group_stats = FALSE,
  expand = TRUE,
  shrink = FALSE,
  schema = "none"
)

record_group(df, ..., to_s4 = TRUE)
}
\arguments{
\item{criteria}{\code{list} of attributes to compare. Each element of the list is a stage in the linkage process. See \code{Details}.}

\item{sub_criteria}{\code{list} of additional attributes to compare at each stage of the linkage process. Comparisons are done as an exact match or with user-defined logical tests. See \code{\link{sub_criteria}}}

\item{sn}{Unique numerical record identifier. Useful for creating familiar group identifiers.}

\item{strata}{Subsets of the dataset. Record groups are created separately for each \code{strata}. Assigning \code{NA} to \code{strata} will skip that record.}

\item{data_source}{Unique data source identifier. Includes a list of data sources of for each record in a \code{pid}.}

\item{data_links}{A set of \code{data_sources} required in a \code{pid}. A \code{pid} without records from these \code{data_sources} are skipped or unlinked. See \code{Details}.}

\item{display}{The messages printed on screen. Options are; \code{"none"} (default) or, \code{"progress"} and \code{"stats"} for a progress update or a more detailed breakdown of the linkage process.}

\item{group_stats}{If \code{TRUE} (default), returns group-specific information like record counts. See \code{Value}.}

\item{expand}{If \code{TRUE}, allows a record group to expand with each subsequent stage of the linkage process.}

\item{shrink}{If \code{TRUE}, forces a record group to shrink with each subsequent stage of the linkage process.}

\item{schema}{Return a schema of the \code{\link[=pid-class]{pid}} object. Options are; \code{"none"} (default), \code{"by_pid"}, \code{"by_strata"} or \code{"by_ALL"}.}

\item{df}{\code{data.frame}. One or more datasets appended together. See \code{Details}.}

\item{...}{Arguments passed to \bold{\code{links}}}

\item{to_s4}{Object type of the returned object. \code{\link[=pid-class]{pid}} (\code{TRUE}) or \code{data.frame} (\code{FALSE}).}
}
\value{
\code{\link[=pid-class]{pid}} or \code{list} (\code{\link[=pid-class]{pid}} and \code{ggplot}) object
}
\description{
Link records with matching criteria in ordered stages of relevance.
Each set of linked records are assigned a unique identifier with relevant group-level data.
}
\details{
Priority of matches in multistage linkages is given to matches that occurred at earlier stages (\code{criteria}) of the linkage process.
Therefore, it's important for each \code{criteria} to be listed in order of decreasing relevance.

Records with missing \code{criteria} (\code{NA}) are skipped at each stage.
Another attempt will be made to match the record at the next stage.
If a record does not match any other record by the end of the linkage process, it is assigned a unique group ID.

A \code{sub_criteria} is a means of specifying additional matching conditions for a stage of the linkage process.
When used, only records with matching \code{criteria} and \code{sub_criteria} are linked.

Each \code{sub_criteria} must be linked to a \code{criteria} in \bold{\code{\link{links}}}.
You can also link multiple \code{sub_criteria} to one \code{criteria}.
Any unlinked \code{sub_criteria} will be ignored.

By default, records are compared for an exact match.
However, user-defined logical tests (function) are also permitted. Such test must meet 3 requirements:
\enumerate{
\item It must be able to compare two atomic vectors
\item It must have two arguments at least two arguments named \code{`x`} and \code{`y`}, where \code{`y`} is the value for one observation being compared against all other observations (\code{`x`}).
\item It must return either TRUE or FALSE.
}

\code{data_links} must be a \code{list} of \code{atomic} vectors, with every element named \code{"l"} (links) or \code{"g"} (groups).

\itemize{
\item if named \code{"l"}, only groups with records from every listed \code{data_source} will be retained.
\item if named \code{"g"}, only groups with records from any listed \code{data_source} will be retained.
}

\bold{\code{record_group()}} as it existed before \code{v0.2.0} has been retired.
Its now exists to support previous code and arguments with minimal disruption. Please use \bold{\code{links()}} instead.

See \code{vignette("links")} for more information.
}
\examples{
library(diyar)
# Exact match
links(criteria = c("Obinna","James","Ojay","James","Obinna"))

# User-defined tests using `sub_criteria()`
# Matching `sex` and + 20-year age gaps
age <- c(30, 28, 40, 25, 25, 29, 27)
sex <- c("M", "M", "M", "F", "M", "M", "F")
f1 <- function(x, y) (y - x) \%in\% 0:20
links(criteria = sex,
      sub_criteria = list(cr1 = sub_criteria(age, funcs = f1)))

# Multistage linkage
# Relevance of matches: `forename` > `surname`
data(staff_records); staff_records
links(criteria = list(staff_records$forename, staff_records$surname),
      data_source = staff_records$sex)

# Relevance of matches:
# `staff_id` > `age` AND (`initials`, `hair_colour` OR `branch_office`)
data(missing_staff_id); missing_staff_id
links(criteria = list(missing_staff_id$staff_id, missing_staff_id$age),
      sub_criteria = list(s2 = sub_criteria(missing_staff_id$initials,
                                          missing_staff_id$hair_colour,
                                          missing_staff_id$branch_office)),
      data_source = missing_staff_id$source_1)

# Group expansion and shrinkage
match_cri <- list(c(1,NA,NA,1,NA,NA),
                  c(1,1,1,2,2,2),
                  c(3,3,3,2,2,2))
links(criteria = match_cri, expand = TRUE)
links(criteria = match_cri, expand = FALSE)
links(criteria = match_cri, shrink =  TRUE)

}
\seealso{
\code{\link{episodes}}, \code{\link{partitions}}, \code{\link{predefined_tests}} and \code{\link{sub_criteria}}
}
