% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_group.R
\name{links}
\alias{links}
\alias{record_group}
\title{Multistage deterministic record linkage}
\usage{
links(criteria, sub_criteria = NULL, sn = NULL, strata = NULL,
  to_s4 = TRUE, data_source = NULL, data_links = "ANY",
  display = "progress")

record_group(df, ...)
}
\arguments{
\item{criteria}{\code{list} of attributes to compare. Comparison is done as an exact match i.e. (\code{==}). See \code{Details}.}

\item{sub_criteria}{\code{list} of additional attributes to compare. Comparison is done as an exact match or a user defined \code{function}. See \code{\link{sub_criteria}}}

\item{sn}{Unique numerical record identifier. Useful for creating familiar episode identifier. Optional.}

\item{strata}{Subsets of the dataset. Links are determined separately within each subset.}

\item{to_s4}{If \code{TRUE} (default), record groups are returned as a \code{\link[=pid-class]{pid}} object.}

\item{data_source}{Unique data source identifier. Useful when the dataset contains data from multiple sources.}

\item{display}{Message printed on screen. Options are; \code{"none"} (default) or, \code{"progress"} and \code{"stats"} for a progress update or a more detailed breakdown of the linkage process.}

\item{df}{\code{data.frame}. One or more datasets appended together. See \code{Details}.}

\item{group_stats}{If \code{TRUE} (default), group specific information like record counts. See \code{Value}.}
}
\value{
\code{\link[=pid-class]{pid}} objects or \code{data.frame} if \code{to_s4} is \code{FALSE})

\itemize{
\item \code{sn} - unique record identifier as provided (or generated)
\item \code{pid | .Data} - unique group identifier
\item \code{link_id} - unique record identifier of matching records
\item \code{pid_cri} - matching criteria
\item \code{pid_dataset} - data sources in each group
\item \code{pid_total} - number of records in each group
}
}
\description{
Link records with matching criteria with different levels of relevance.
}
\details{
\bold{\code{links()}} performs an ordered multi-staged deterministic linkage.
The order (relevance/priority) of this is determined by the order of in which each \code{criteria} is listed.

\code{sub_criteria} specifies additional matching conditions for each stage (\code{criteria}) of the process.
If \code{sub_criteria} is not \code{NULL}, only records with a matching \code{criteria} and \code{sub_criteria} are grouped together.
If a record has missing values for any \code{criteria}, that record is skipped at that stage, and another attempt is made at the next stage.
If there are no matches for a record at every stage, that record is assigned a unique group ID.

By default, records are compared as an exact match i.e. the equivalent of \code{(==)} however, user defined functions are also permitted.
The function must be able to compare two atomic vectors,
return only \code{TRUE} or \code{FALSE},
and have two arguments - \code{x} for the attribute and \code{y} for what it'll be compared against.

A match at each stage is considered more relevant than a match at the next stage. Therefore, \code{criteria} should always be listed in order of decreasing relevance.

\code{data_source} - including this returns \code{pid_dataset}. This lists the source of every event in each record group.

\bold{\code{record_group()}} as it existed before \code{v0.2.0} has been retired.
Its current implementation only exists to support existing code with minimal disruption. Please use \bold{\code{links()}} moving forward.
}
\examples{
library(diyar)
# Exact match
links(criteria = c("Obinna","James","Ojay","James","Obinna"))

# Use defined tests - `sub_criteria()`
# Matching `sex` an + 20-year age gaps
age <- c(30,28,40,25,25,29,27)
sex <- c("M","M","M","F","M","M","F")
f1 <- function(x, y) (y - x) \%in\% 0:20
links(criteria = sex,
      sub_criteria = list(s1 = sub_criteria(dob$age, funcs = f1)))

# Matching `sex` an +/- 20-year age gaps
f2 <- function(x, y) abs(y - x) \%in\% abs(0:20)
links(criteria = sex,
      sub_criteria = list(s1 = sub_criteria(dob$age, funcs = f2)))

# Multistage linkage
# Relevance of matches: `forename` > `surname`
data(staff_records); staff_records
links(staff_records,
      criteria = list(staff_records$forename, staff_records$surname),
      data_source = staff_records$sex)

# Relevance of matches:
# `staff_id` > `age` AND (`initials`, `hair_colour` or `branch_office`)
data(missing_staff_id); missing_staff_id
links(criteria = list(missing_staff_id$staff_id, missing_staff_id$age),
      sub_criteria = list(s2=sub_criteria(missing_staff_id$initials,
                                          missing_staff_id$hair_colour,
                                          missing_staff_id$branch_office)),
      data_source = missing_staff_id$source_1)
}
\seealso{
\code{\link{episodes}} and \code{\link{number_line}}
}
