% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{number_line-class}
\alias{number_line-class}
\alias{show,number_line-method}
\alias{rep,number_line-method}
\alias{[,number_line-method}
\alias{[[,number_line-method}
\alias{[<-,number_line-method}
\alias{[[<-,number_line-method}
\alias{$,number_line-method}
\alias{$<-,number_line-method}
\alias{c,number_line-method}
\alias{unique.number_line}
\alias{sort.number_line}
\alias{format.number_line}
\title{\code{number_line} object}
\usage{
\S4method{show}{number_line}(object)

\S4method{rep}{number_line}(x, ...)

\S4method{[}{number_line}(x, i, j, ..., drop = TRUE)

\S4method{[[}{number_line}(x, i, j, ..., exact = TRUE)

\S4method{[}{number_line}(x, i, j, ...) <- value

\S4method{[[}{number_line}(x, i, j, ...) <- value

\S4method{$}{number_line}(x, name)

\S4method{$}{number_line}(x, name) <- value

\S4method{c}{number_line}(x, ...)

\method{unique}{number_line}(x, ...)

\method{sort}{number_line}(x, decreasing = FALSE, ...)

\method{format}{number_line}(x, ...)
}
\arguments{
\item{object}{object}

\item{x}{x}

\item{...}{...}

\item{i}{i}

\item{j}{j}

\item{drop}{drop}

\item{exact}{exact}

\item{value}{value}

\item{name}{slot name}

\item{decreasing}{logical. Should the sort be increasing or decreasing}
}
\description{
S4 objects representing a series of finite numbers on a number line
Used for range matching in \code{\link{record_group}} and interval grouping in \code{\link{fixed_episodes}}, \code{\link{rolling_episodes}} and \code{\link{episode_group}}
}
\section{Slots}{

\describe{
\item{\code{start}}{Start of the number line}

\item{\code{id}}{Unique \code{numeric} ID. Providing this is optional.}

\item{\code{gid}}{Unique \code{numeric} Group ID. Providing this is optional.}

\item{\code{.Data}}{Length/with and direction of the \code{number_line} object.}
}}

