% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_line.R
\docType{class}
\name{number_line-class}
\alias{number_line-class}
\title{Number line objects
S4 objects representing a series of finite numbers on a number line
Used for range matching in \code{\link{record_grouping}} and interval grouping in \code{\link{episode_grouping}}}
\description{
Number line objects
S4 objects representing a series of finite numbers on a number line
Used for range matching in \code{\link{record_grouping}} and interval grouping in \code{\link{episode_grouping}}
}
\section{Slots}{

\describe{
\item{\code{start}}{Start of the number line}

\item{\code{id}}{Unique \code{numeric} ID. Providing this is optional.}

\item{\code{.Data}}{Length/with and direction of the number line}
}}

