% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_line.R
\docType{class}
\name{number_line-class}
\alias{number_line-class}
\alias{number_line_obj}
\title{Number lines
S4 objects representing a number line.
Used for range matching in [record_grouping] and interval grouping in [episode_grouping].}
\description{
Number lines
S4 objects representing a number line.
Used for range matching in [record_grouping] and interval grouping in [episode_grouping].
}
\section{Slots}{

\describe{
\item{\code{start}}{Start of the number line.}

\item{\code{.Data}}{Length or width of the number line.}
}}

