% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_line.R
\name{number_line}
\alias{number_line}
\alias{is.number_line}
\title{Number line}
\usage{
number_line(a, z)

is.number_line(x)
}
\arguments{
\item{a}{Start of the number line. Should be a vector of \code{numeric}, \code{double} or \code{integer}}

\item{z}{End of the number line. Should be a vector of \code{numeric}, \code{double} or \code{integer}}
}
\value{
\code{number_line} object
}
\description{
A function to create number lines
}
\examples{

number_line(-100, 100)
number_line(10, 11.2)
number_line(100, 200)
number_line(100, 120)
number_line(50, 120)
number_line(100,100)
}
