% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_line.R
\name{number_line}
\alias{number_line}
\alias{is.number_line}
\alias{reverse}
\title{Number line}
\usage{
number_line(a, z)

is.number_line(x)

reverse(x, direction = "both")
}
\arguments{
\item{a}{Start of the number line. It should a \code{numeric} object, or can be coerced to a \code{numeric} object}

\item{z}{End of the number line. It should a \code{numeric} object, or can be coerced to a \code{numeric} object}

\item{x}{object}

\item{x}{object}
}
\value{
\code{number_line} object
}
\description{
A set of function to create and manipulate \code{number_line} objects
}
\examples{
library(lubridate)

number_line(-100, 100)
number_line(10, 11.2)
number_line(dmy("12/03/2019"), dmy("05/01/2019"))
number_line(dmy_hms("15/05/2019 13:15:07"), dmy_hms("15/05/2019 15:17:10"))
a <- number_line(0, -100)
b <- number_line(dmy("25/04/2019"), dmy("01/01/2019"))
is.number_line(a)
is.number_line(b)


reverse(number_line(dmy("25/04/2019"), dmy("01/01/2019")))
reverse(number_line(200,-100), "increasing")
reverse(number_line(200,-100), "decreasing")

}
