% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlaps}
\alias{overlaps}
\alias{overlap}
\alias{exact}
\alias{across}
\alias{chain}
\alias{aligns_start}
\alias{aligns_end}
\alias{inbetween}
\alias{overlap_method}
\alias{include_overlap_method}
\alias{exclude_overlap_method}
\title{Overlapping number line objects}
\usage{
overlaps(x, y, methods = "overlap", method = c("exact", "across",
  "chain", "aligns_start", "aligns_end", "inbetween", "overlap", "none"))

overlap(x, y)

exact(x, y)

across(x, y)

chain(x, y)

aligns_start(x, y)

aligns_end(x, y)

inbetween(x, y)

overlap_method(x, y)

include_overlap_method(methods)

exclude_overlap_method(methods)
}
\arguments{
\item{x}{\code{number_line} object}

\item{y}{\code{number_line} object}

\item{methods}{Methods of overlap. Check different pairs of \code{number_line} objects using different \code{methods}}

\item{method}{Method of overlap. Check every pair of \code{number_line} objects with the same \code{method}. Deprecated. Please use \code{methods} instead.}
}
\value{
\code{logical}; \code{character}
}
\description{
Identify overlapping \code{number_line} objects
}
\details{
\bold{7 logical test;}

\bold{\code{exact()}} - Identical start and end points

\bold{\code{inbetween()}} - start and end points of one \code{number_line} object is in between the start and end points of another.

\bold{\code{across()}} - Start or end points of one \code{number_line} object is in between the start and end points of another.

\bold{\code{chain()}} - Chained i.e. end point of one \code{number_line} object is the same as the start point of another.

\bold{\code{aligns_start()}} - Identical start points only.

\bold{\code{aligns_end()}} - Identical end points only.

\bold{\code{overlap()}} - Any kind of overlap.
All other methods are mutually exclusive.
\bold{\code{overlap()}} is just a convenient \code{method} for "ANY" and "ALL" methods of overlap.

\bold{\code{overlaps()}} - Overlap by any or all 7 methods above.

\bold{Describe methods of overlap;}

\bold{\code{overlap_method()}} - Shows if and how a pair of \code{number_line} object has overlapped.
Does not show \code{"overlap"} since \bold{\code{overlap()}} is always \code{TRUE} when any other method is \code{TRUE}.

\bold{\code{include_overlap_method()}} and \bold{\code{exclude_overlap_method()}} - Conveniently create the required values for \code{methods} and \code{overlap_methods} in \code{\link{episode_group}}.
}
\examples{
a <- number_line(-100, 100)
b <- number_line(10, 11.2)
c <- number_line(100, 200)
d <- number_line(100, 120)
e <- number_line(50, 120)
g <- number_line(100,100)

overlaps(a, g)
overlaps(a, g, methods = "exact|chain")

overlap(a, b)
overlap(a, e)

exact(a, g)
exact(a, a)

across(a, b)
across(a, e)

chain(c, d)
chain(a, c)

aligns_start(c, d)
aligns_start(a, c)

aligns_end(d, e)
aligns_end(a, c)

inbetween(a, g)
inbetween(b, a)

overlap_method(a, c)
overlap_method(d, c)
overlap_method(a, g)
overlap_method(b, e)

include_overlap_method("across")
include_overlap_method(c("across", "chain"))

exclude_overlap_method("across")
exclude_overlap_method(c("across", "chain"))
}
\seealso{
\code{\link{number_line}} and \code{\link{set_operations}}
}
