% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{staff_records}
\alias{staff_records}
\alias{missing_staff_id}
\alias{infections}
\alias{infections_2}
\alias{infections_3}
\alias{infections_4}
\alias{hospital_admissions}
\alias{patient_list}
\alias{patient_list_2}
\alias{hourly_data}
\alias{Opes}
\alias{episode_unit}
\title{Datasets in \code{diyar} package}
\format{tibble}
\usage{
data(staff_records)

data(missing_staff_id)

data(infections)

data(infections_2)

data(infections_3)

data(infections_4)

data(hospital_admissions)

data(patient_list)

data(patient_list_2)

data(hourly_data)

data(Opes)

data(episode_unit)
}
\description{
Datasets in \code{diyar} package
}
\details{
staff_records - Staff record with some missing data

missing_staff_id - Staff records with missing staff identifiers

infections, infections_2, infections_3 and infections_4 - Reports of bacterial infections

hospital_admissions - Hospital admissions and discharges

patient_list & patient_list_2 - Patient list with some missing data

Hourly data

Opes - List of individuals with the same name

Duration in seconds for each 'episode_unit'
}
\examples{
data(staff_records)
data(missing_staff_id)
data(infections)
data(infections_2)
data(infections_3)
data(infections_4)
data(hospital_admissions)
data(patient_list)
data(patient_list_2)
data(hourly_data)
data(Opes)
data(episode_unit)
}
\keyword{datasets}
