% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_group.R
\name{record_group}
\alias{record_group}
\title{Deterministic linkage for epidemiological analysis}
\usage{
record_group(df, sn, criteria, sub_criteria = NULL, data_source = NULL,
  display = TRUE)
}
\arguments{
\item{df}{Dataframe. One or more datasets appened together.}

\item{sn}{Unique record indentifier for the dataframe.}

\item{criteria}{Matching criteria. Records with matching values in these columns are grouped together. Records are grouped in order of decreasing certainty. For exampe, the first criteria is more certian than second therefore, records matching on the first criteria will be grouped together in preference to those matching on the second criteria.}

\item{sub_criteria}{Matching sub-criteria. If is not \code{NULL}, only records with matching values in the \code{criteria} and \code{sub_criteria} columns are grouped together. Usefull in providing alternative or less stringent matching criteria.}

\item{data_source}{Unique dataset indentifier for the dataframe. Usefull when dataframe contains multiple datsets.}

\item{display}{If \code{TRUE}, progress status at each stage of record grouping is displayed on screen.}
}
\value{
Dataframe
}
\description{
This function links records from one or more datasets into a unique idenfier. Linkage is based on specified matching criteria.
}
\examples{


}
