% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_group.R
\name{record_group}
\alias{record_group}
\title{Deterministic linkage for epidemiological analysis}
\usage{
record_group(df, sn, criteria, sub_criteria = NULL, display = TRUE)
}
\arguments{
\item{df}{A dataframe}

\item{sn}{A column of unique indentifiers for each record in 'df'}

\item{criteria}{Character vector of columns in 'df' representing matching criteria}

\item{sub_criteria}{Character vector of columns in 'df' representing matching sub-criteria within each 'criteria'}
}
\value{
Dataframe with a unique group identifier based on the matching criteria and sub-criter
}
\description{
This function links records from one ore more datasets on a unique idenfier using matching or partial matching criteria
}
\examples{



}
