% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reframe.R
\name{reframe}
\alias{reframe}
\alias{reframe.sub_criteria}
\title{Modify \code{sub_criteria} objects}
\usage{
reframe(x, ...)

\method{reframe}{sub_criteria}(x, func = identity, ...)
}
\arguments{
\item{x}{\code{[\link{sub_criteria}]}.}

\item{...}{Arguments passed to methods.}

\item{func}{\code{[function]}. Transformation function.}
}
\description{
Modify the attributes of a \code{\link{sub_criteria}} object.
}
\examples{
s_cri <- sub_criteria(month.abb, month.name)
reframe(s_cri, func = function(x) x[12])
reframe(s_cri, func = function(x) x[12:1])
reframe(s_cri, func = function(x) attrs(x[1:6], x[7:12]))
}
\seealso{
\code{\link{sub_criteria}}; \code{\link{eval_sub_criteria}}; \code{\link{attr_eval}}
}
