% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{schema}
\alias{schema}
\alias{schema.pane}
\alias{schema.epid}
\alias{schema.pid}
\title{schema}
\usage{
schema(x, ...)

\method{schema}{pane}(x, int, separate, title = NULL, show_labels = FALSE)

\method{schema}{epid}(
  x,
  date,
  episode_unit,
  case_length,
  recurrence_length = case_length,
  episode_type,
  from_last,
  title = NULL,
  show_labels = FALSE,
  show_skipped = TRUE,
  show_non_finite = FALSE
)

\method{schema}{pid}(x, title = NULL, show_labels = FALSE)
}
\arguments{
\item{title}{Title of \code{ggplot} object}

\item{pids}{\code{pid} object}
}
\value{
\code{ggplot}
}
\description{
schema
}
\details{

}
