% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema}
\alias{schema}
\alias{schema.number_line}
\alias{schema.epid}
\alias{schema.pane}
\alias{schema.pid}
\title{Schema diagram for linked records in \code{diyar}}
\usage{
schema(x, ...)

\method{schema}{number_line}(x, show_labels = c("date", "case_overlap_methods"), ...)

\method{schema}{epid}(
  x,
  title = NULL,
  show_labels = c("length_arrow"),
  show_skipped = TRUE,
  show_non_finite = FALSE,
  theme = "dark",
  seed = NULL,
  custom_label = NULL,
  ...
)

\method{schema}{pane}(
  x,
  title = NULL,
  show_labels = c("window_label"),
  theme = "dark",
  seed = NULL,
  custom_label = NULL,
  ...
)

\method{schema}{pid}(
  x,
  title = NULL,
  show_labels = TRUE,
  theme = "dark",
  orientation = "by_pid",
  seed = NULL,
  custom_label = NULL,
  ...
)
}
\arguments{
\item{x}{\code{[\link[=number_line-class]{number_line}|\link[=pid-class]{pid}|\link[=epid-class]{epid}|\link[=pane-class]{pane}]}}

\item{...}{Other arguments.}

\item{show_labels}{\code{[logical|character]}. Show/hide certain parts of the schema. See \code{Details}.}

\item{title}{\code{[character]}. Plot title.}

\item{show_skipped}{\code{[logical]}. Show/hide \code{"Skipped"} records.}

\item{show_non_finite}{\code{[logical]}. Show/hide records with non-finite \code{date} values.}

\item{theme}{\code{[character]}. Options are \code{"dark"} or \code{"light"}.}

\item{seed}{\code{[integer]}. See \code{set.seed}. Used to get a consistent arrangement of items in the plot.}

\item{orientation}{\code{[character]}. Show each record of a \code{pid} object within its group id (\code{"by_pid"}) or its \code{pid_cri} (\code{"by_pid_cri"})}
}
\value{
\code{ggplot} objects
}
\description{
Create schema diagrams for \code{pid}, \code{epid} and \code{pane} objects.
}
\details{
The \code{schema} function creates a schema diagram for \code{\link[=number_line-class]{number_line}}, \code{\link[=pid-class]{pid}}, \code{\link[=epid-class]{epid}} and \code{\link[=pane-class]{pane}} objects.
These diagrams are a visual aid to describing the data linkage (\code{\link{links}}), episode tracking (\code{\link{episodes}}) or partitioning process (\code{\link{partitions}}).

\emph{\code{\link[=number_line-class]{number_line}} objects are plotted as \code{\link[=epid-class]{epid}} objects.}
}
\examples{
schema(number_line(c(1, 2, 6, 8), c(5, 3, 8, 10)))

schema(episodes(1:10, 2))

schema(partitions(1:10, by = 2, separate = TRUE))

schema(links(list(c(1, 1, NA, NA), c(NA, 1, 1, NA))))

}
