% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema}
\alias{schema}
\alias{schema.pane}
\alias{schema.epid}
\alias{schema.pid}
\title{Schema  diagram for linked records.}
\usage{
schema(x, ...)

\method{schema}{pane}(x, title = NULL, show_labels = c("window_label"), dark_mode = TRUE, ...)

\method{schema}{epid}(
  x,
  title = NULL,
  show_labels = c("length_arrow"),
  show_skipped = TRUE,
  show_non_finite = FALSE,
  dark_mode = TRUE,
  ...
)

\method{schema}{pid}(
  x,
  title = NULL,
  show_labels = TRUE,
  dark_mode = TRUE,
  orientation = "by_pid",
  ...
)
}
\arguments{
\item{x}{\code{pid}, \code{epid}, or \code{pane} object.}

\item{...}{XXXX}

\item{title}{Title of \code{ggplot} object}

\item{show_labels}{Show/hide certain parts of the schema. See \code{Details}}

\item{dark_mode}{Use a black/white background for each plot.}

\item{show_skipped}{Show/hide \code{"Skipped"} records}

\item{show_non_finite}{Show/hide records with non-finite \code{date} values}

\item{orientation}{Show each record of a \code{pid} object within its group id (\code{"by_pid"}) or its \code{pid_cri} (\code{"by_pid_cri"})}
}
\value{
list of \code{ggplot} objects
}
\description{
Create schema diagrams for \code{pid}, \code{epid} and \code{pane} objects.
}
\details{
The \code{schema} function creates a schema diagram for \code{pid}, \code{epid} and \code{pane} objects.
These diagrams are a visual aid to describing the data linkage (\code{\link{links}}), episode tracking (\code{\link{episodes}}) or partitioning process(\code{\link{partitions}}).

When used for \code{epid} and \code{pane} objects, additional information in the form of arguments passed to the instance \code{\link{episodes}} and \code{\link{partitions}} will be required.
}
