% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_group.R
\name{sub_criteria}
\alias{sub_criteria}
\title{Subcriteria for \bold{\code{\link{links}}}}
\usage{
sub_criteria(..., funcs = NULL)
}
\arguments{
\item{...}{Additional attributes to compare.}

\item{funcs}{User defined logical test.}
}
\value{
\code{list}
}
\description{
Additional \code{sub_criteria} for each matching \code{criteria} in \code{\link{links}}.
}
\details{
\bold{\code{sub_criteria()}} is the mechanism for providing a subcriteria to an instance of \bold{\code{links()}}.

Each attribute (\code{atomic}) is compared as an exact match or with a user defined logical test.
The test must be supplied as a funciton with at least two arguments named \code{`x`} and \code{`y`},
where \code{`y`} is the attribute for one observation being compared against all other attributes (\code{`x`}).

Each attribute must have the same length.

\code{funcs} must be a \code{function} or \code{list} of \code{functions} to use on each attribute.

Each \code{funcs} must evaluate to \code{TRUE} or \code{FALSE}.

Each \code{sub_criteria} must be linked to a \code{criteria} in \bold{\code{\link{links}}}.
You can link mutliple \code{sub_criteria} to one \code{criteria}.
It's not neccessary to link a \code{sub_criteria} if there's only one \code{criteria}.
Any unlinked \code{sub_criteria} will be ignored.
}
\examples{
library(diyar)

sub_criteria(c(30, 28, 40, 25, 25, 29, 27), funcs = range_matching)

# Link each `sub_critera` a `criteria`.
list(
c1 = sub_criteria(c(30, 28, 40, 25, 25, 29, 27), funcs = range_matching),
c2 = sub_criteria(c(30, 28, 40, 25, 25, 29, 27), funcs = range_matching))
}
