% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_group.R
\name{sub_criteria}
\alias{sub_criteria}
\title{Sub-criteria}
\usage{
sub_criteria(..., funcs = diyar::exact_match, equva = diyar::exact_match)
}
\arguments{
\item{...}{Additional attributes to compare.}

\item{funcs}{User defined logical test.}
}
\value{
\code{list}
}
\description{
Additional matching conditions when using \bold{\code{\link{links}}} and \bold{\code{\link{episodes}}}
}
\details{
\bold{\code{sub_criteria()}} is the mechanism for providing a \code{sub_criteria} to an instance of \bold{\code{links}} or \bold{\code{episodes}}.

Each attribute (\code{atomic} vectors) is compared with a user-defined logical test.

Each attribute must have the same length.

\code{funcs} must be a \code{function} or \code{list} of \code{functions} to use with each attribute.

Each \code{funcs} must evaluate to \code{TRUE} or \code{FALSE}.
}
\examples{
library(diyar)

sub_criteria(c(30, 28, 40, 25, 25, 29, 27), funcs = range_match)

# Link each `sub_critera` a `criteria`.
list(
c1 = sub_criteria(c(30, 28, 40, 25, 25, 29, 27), funcs = range_match),
c2 = sub_criteria(c(30, 28, 40, 25, 25, 29, 27), funcs = range_match))
}
