% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_s4.R
\name{to_df}
\alias{to_df}
\alias{to_s4}
\title{Change S4 objects in \code{diyar} to data frames and vice versa}
\usage{
to_df(s4, ...)

to_s4(lst)
}
\arguments{
\item{s4}{\code{[\link[=pid-class]{pid}|\link[=epid-class]{epid}|\link[=pane-class]{pane}|\link[=number_line-class]{number_line}]}}

\item{...}{Arguments passed to \code{data.frame}}

\item{lst}{\code{[list]}}
}
\value{
to_df - \code{data.frame} object

to_s4 - \link[=pid-class]{pid}, \link[=epid-class]{epid}, \link[=pane-class]{pane} or \link[=number_line-class]{number_line} objects
}
\description{
Convert \link[=pid-class]{pid}, \link[=epid-class]{epid}, \link[=pane-class]{pane} or \link[=number_line-class]{number_line} objects
to a \code{data.frame} and vice versa.
}
\details{
\code{to_df} has been retired. Moving forward, please use \code{as.data.frame}.
}
