% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_s4.R
\name{to_df}
\alias{to_df}
\alias{to_s4}
\title{Change S4 objects in \code{diyar} to data frames and vice versa}
\usage{
to_df(s4, ...)

to_s4(df)
}
\arguments{
\item{s4}{\code{[\link[=pid-class]{pid}|\link[=epid-class]{epid}|\link[=pane-class]{pane}|\link[=number_line-class]{number_line}]}}

\item{...}{Arguments passed to \code{data.frame}}

\item{df}{\code{[data.frame]}}
}
\value{
to_df - \code{data.frame} object

to_s4 - \link[=pid-class]{pid}, \link[=epid-class]{epid}, \link[=pane-class]{pane} or \link[=number_line-class]{number_line} objects
}
\description{
Convert \link[=pid-class]{pid}, \link[=epid-class]{epid}, \link[=pane-class]{pane} or \link[=number_line-class]{number_line} objects
to a \code{data.frame} and vice versa.
}
\details{
\code{to_df} has been retired. Moving forward, please use \code{as.data.frame}.
}
\examples{
data(infections)
dates <- infections$date
output <- episodes(dates, case_length=30)
output

# Change a pid, epid, pane or number_line object to a data.frame
df_output <- to_df(output)
df_output

# Change a data.frame to a pid, epid, pane or number_line object
s4_output <- to_s4(df_output)
s4_output

}
